pyproject.toml
[project]
name = "autokupon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "blinker==1.7.0",
    "packaging>=25.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.39.0",
    "selenium-wire>=5.1.0",
    "setuptools>=80.9.0",
]
